feat(ui): code refactoring
