+++ apps/capital-azur/theme/card.js
@@ -45,7 +45,7 @@ export const card = {
 		image: "flex-shrink-0",
 		body: "px-4 py-3",
 		tag: "text-sm font-medium text-indigo-600",
-		title: "text-xl font-semibold text-gray-900 text-black hover:text-indigo-600",
+		title: "text-xl font-semibold text-gray-900 text-black hover:text-indigo-600 mb-5",
 		excerpt: "mt-3 text-base text-gray-500 dark:text-gray-400",
 		link: "text-primary underline",
 	},
